requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
click==7.1.2
Django==3.0.6
django-cors-headers==3.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
pathspec==0.8.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
regex==2020.5.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
